changed days to get data from server in planForm
