+++ src/assets/css/global.css
@@ -29,3 +29,9 @@ body,
 .alert-exit {
   opacity: 1;
 }
+
+.alert-exit-active {
+  opacity: 0;
+  transform: scale(0.9);
+  transition: opacity 300ms, transform 300ms;
+}
